// packages/frontend/src/api/client/models/TaskUpdateDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TaskUpdateDto = {
    /**
     * Task name
     */
    name?: string;
    /**
     * Optional task description
     */
    description?: Record<string, any>;
    /**
     * Prompt/template used for the task
     */
    prompt?: string;
};